Fixes sudden wizard and simplifies root init logic
- The issue with the sudden wizard was that during the init process,
the app looks for an active character, and if it can't find one, then it
shows the wizard. It used to be that the active character method would return
the default character if no active one was found. This was removed a while back
but without a default character, the app would always default to the wizard.

- The other part of this commit simplifies the init logic. I've removed the
confusing part of the init function which simplified it greatly. init.js is now
soley responsible for setting up notifications and initing the root vm. Inside
the root vm, is the logic for selecting a character and calling load.

Hopefully this makes everything far more apparent.