Adds keyboard shortcut toggle to Actions Toolbar
- During development there was an issue with activating the toolbar via signals
  and hotkeys. This was due to the fact that the Actions Toolbar was attempting
to set the hotkeys during the load/init process. According to the debug logs it
appeared that the call was being made but on a different object than the one
used later in the process. The actual issue turned out to be that the Actions
Toolbar was appearing to call the load/init functions but wasn't as they were
never called in the root.js. I've added the calls to the root.js and it works
fine now.