Adds Dependency Tracking to nested model relationships + Other Fixes
There seems to be a common paradigm of using

    list.forEach( async () => { ... }

to update various values. This is **incorrect** as forEach will not wait
for those promises to be fulfilled before moving forward.

The correct process for this involves using the new-style for-of loop
or a map method that resolves the promises.

    for (const item of list) { await ... }

Source:
https://stackoverflow.com/questions/37576685/using-async-await-with-a-foreach-loop

I've also fixed an issue that would cause the Health to force the death
saves to reload when their values didn't change.