Adds new POI DC check fields.
These fields are optional and do not block submissions nor apply to old
POIs.

This commit also adds some new flex utilities present in Bootstrap 4/5
but not in 3 and they're really nice.

This commit also redoes the Wealth Quick Spend to use these new
utilities.

POI DC checks can be set to any text and number value, but the skill
autopopulates the basic skills and stats.