Fixes up Navbar display on small screens and cleans up code (#2055)
The navbar used to be kinda broken on small devices and would overflow onto two lines. This is fixed.

Also our navbar would collapse on small devices, but the dropdown would still exist. This essentially gave us two collapsed navbars, one inside of the other. This was very bad for UX. This is also fixed.