Refactors remaining DM tools
The Overview tab has been removed and the ordering of the tabs has changed. Encounters is now first and the new Maps/Images tab is second.

The Profile has also been rewitten to support the same flip-card portrait as the character-side does.

Unneeded code is removed

Minor styling changes and other fixes to other modules in the DM tools and party tools for consistency sake