Picker Improvements + Patron of the Day (#2074)
* First pass

* Adds online status/filters to the party dashboard

* Presence updates now preserve other presence data

* Adds Authentication to WebSocket connection URL

* Cleanup

* Tunes up home page picker and root elements to make things cleaner and better on mobile.

* Finishes Party Service Implementation.

The party service no longer polls frequently for changes. It still polls once every 5 minutes just in case a change is not caught by the awareness updates (i.e. in case we forget something).

The party service also now notifies based on its connection status.

The party service also forces a refresh whenever the awareness changes and then alerts subscribers of that update. This should drastically reduce the need to query the server.

* Presence updates will now always trigger refreshes.

A random number is always inserted into the presence data when manually calling `updatePresence` this is to force the presence update to always occur even when updating the data to exactly the same value as before.

* All DM exhibit updates (and profile changes) cause party update

* Adds the ability to simply touch the presence to update it.

* Character Sheets now send Presence Event Notifications.

This commit adds the ability for player sheets to notify their party of relevant changes to their data. This should work for everything relevant to a character and their data for the player-cards on both sides.

I have also tweaked the debounce timers to enable them to better coallate the events to prevent us from causing repeated presences from spamming our database. These debounces coallate events both before they are sent and before recieved events trigger a refresh from the server.

This commit also addresses some UI differences between the Player/DM party dashboards. Both can now filter by "online status".

* Party Service now accounts for disconnecting and switching cores and better accounts for a player not connected to a party.

This commit ensures better safety for other VMs by ensuring it is always safe to call `updatePresence` even if not connected to a party. In that case it is simply a no-op.

* Adds support for the new proxy config

* Initial work on patron of the day

* Adds Patron of the Day to root picker screen

* Better wording on patron of the day