Adds Initiative Tracker Tab for DMs
This commit adds a new tab that contains an Initiative Tracker tool that can be used to track the current status of any in-game encounter in turns. It tracks time as well as rounds, and automatically includes players from the party in the list.

Additionally, monsters from the Monster Encounter section can be easily added to the tracker.

The tracker also uses a dice roller to automatically calculate initiative and allows for the DM to override these values if needed.